// src/app/heroes/interfaces/hero.interface.ts
// Generated by https://quicktype.io
export interface CountAndHero {
  total:number;
  heroes:Hero[]
}

export interface Hero {
  slug:             string;
  superhero:        string;
  publisher:        Publisher;
  alter_ego:        string;
  first_appearance: string;
  characters:       string;
  alt_img?:         string;
}

export enum Publisher {
  DCComics = "DC Comics",
  MarvelComics = "Marvel Comics",
}